extract Enum class
