+++ backend/src/main/java/com/ai_kids_care/v1/dto/TokenResponse.java
@@ -1,9 +1,8 @@
 package com.ai_kids_care.v1.dto;
 
+import com.ai_kids_care.v1.type.TokenTypeEnum;
 import lombok.*;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 
@@ -19,64 +18,26 @@ import jakarta.annotation.Generated;
 @Builder
 public class TokenResponse {
 
-  
   @NonNull
   @Schema(name = "accessToken", requiredMode = Schema.RequiredMode.REQUIRED)
   @JsonProperty("accessToken")
   private String accessToken;
 
-  /**
-   * Gets or Sets tokenType
-   */
-  public enum TokenTypeEnum {
-    BEARER("Bearer");
-
-    private final String value;
-
-    TokenTypeEnum(String value) {
-      this.value = value;
-    }
-
-    @JsonValue
-    public String getValue() {
-      return value;
-    }
-
-    @Override
-    public String toString() {
-      return String.valueOf(value);
-    }
-
-    @JsonCreator
-    public static TokenTypeEnum fromValue(String value) {
-      for (TokenTypeEnum b : TokenTypeEnum.values()) {
-        if (b.value.equals(value)) {
-          return b;
-        }
-      }
-      throw new IllegalArgumentException("Unexpected value '" + value + "'");
-    }
-  }
-
-  
   @NonNull
   @Schema(name = "tokenType", requiredMode = Schema.RequiredMode.REQUIRED)
   @JsonProperty("tokenType")
   private TokenTypeEnum tokenType;
 
-  
   @NonNull
   @Schema(name = "expiresIn", description = "seconds", requiredMode = Schema.RequiredMode.REQUIRED)
   @JsonProperty("expiresIn")
   private Integer expiresIn;
 
-  
   @NonNull
   @Schema(name = "refreshToken", requiredMode = Schema.RequiredMode.REQUIRED)
   @JsonProperty("refreshToken")
   private String refreshToken;
 
-  
   @Schema(name = "refreshExpiresIn", description = "seconds", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
   @JsonProperty("refreshExpiresIn")
   private Integer refreshExpiresIn = null;
