+++ backend/src/main/java/com/ai_kids_care/v1/type/TokenTypeEnum.java
@@ -1,0 +1,37 @@
+package com.ai_kids_care.v1.type;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonValue;
+
+/**
+ * Gets or Sets tokenType
+ */
+public enum TokenTypeEnum {
+    BEARER("Bearer");
+
+    private final String value;
+
+    TokenTypeEnum(String value) {
+        this.value = value;
+    }
+
+    @JsonValue
+    public String getValue() {
+        return value;
+    }
+
+    @Override
+    public String toString() {
+        return String.valueOf(value);
+    }
+
+    @JsonCreator
+    public static TokenTypeEnum fromValue(String value) {
+        for (TokenTypeEnum b : TokenTypeEnum.values()) {
+            if (b.value.equals(value)) {
+                return b;
+            }
+        }
+        throw new IllegalArgumentException("Unexpected value '" + value + "'");
+    }
+}
